This commit message applied to the previous comment, which was accidentally sent off without a commit message.
Added the copy-on-write pattern into BucketPage and DirectoryPage.  Note that addLevel() really should be moved onto BucketPage in keeping with this pattern. TestCommit is now passing.  All tests which were passing are still passing.

Bug fix to the DirectoryPage deserialization constructor.  It was failing to set the address on the deserialized page.

Added an integration test with the MemoryManager using MemStore.  This demonstrates some problems with the incremental eviction which I have not yet debugged.  Many of the tests in TestHTreeWithMemStore will pass if you increase the write retention queue capacity, but that is because incremental evictions are not occurring.  As an aid to demonstrating problems, I reduced the write retention queue for this test from the default (500) to 10.

Modified SimpleRabaCoder to track the IRaba capacity as well as its size.  The old serialized version assumed that capacity == size on decode, but that was breaking some assumptions with regard to the bucket page capacity.  (We might be able to work around those assumptions in the future if we keep the bucket page dense, but that is not currently guaranteed.)

Removed the invocation of tree.touch() in Leaf/BucketPage prior to the invocation of the copy-on-write pattern.  copyOnWrite() was invoking touch() so this was just extra work and touch() is a hot spot.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@4943 8f7bc0f5-282e-0410-95e3-8d296e9bb460