Bug fix to HTree.lookupAll(), which had not been modified yet to have a single "buddy" for a bucket page.
Added IRawRecordAccess, which is implemented by Leaf and BucketPage.  This interface provides access via the owning index object (B+Tree or HTree) to a raw record via its storage address.  Modified AbstractTuple to use this new interface such that we can resolve raw records on ITupleIterator patterns for the HTree as well as the B+Tree.

Added a new DISTINCT operator based on the HTree with a test suite.  This implementation very closely parallels the existing DISTINCT operator based on a ConcurrentHashMap.  The new DISTINCT operator passes its basic test suite.  

The HTree for the new DISTINCT operator needs to be associated with an IMemoryManager which which is a child of the IMemoryManager for the IRunningQuery.  Currently, it is backed by the ISimpleMemoryRawStore, which is not scalable.  This API change should be easy to introduce as we only need to modify PipelineOp.newStats() to accept the IRunningQuery.

The new DISTINCT operator is not safe for concurrent evaluation.  I need to go back and check the PipelineOp annotations on the operator and verify that it will not be run against concurrent chunks.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@4950 8f7bc0f5-282e-0410-95e3-8d296e9bb460