Checkpoint on aggregation support [1].
Wrote a pipelined aggregation operator (PipelinedAggregationOp).  Refactored the test suite to run against both the MemoryGroupByOp and the new PipelinedAggregationOp.  Extended GroupByState to report when nested aggregates are used and GroupByOp to report if the operator supports pipelined evaluation, in which case it can not be used with DISTINCT or nested aggregates.  Added new unit tests for aggregation when nested aggregates are used. Modified the IAggregate functions to have "sticky" errors which are cleared by reset() and updates the test suites for the IAggregate functions to verify this behavior.

There are currently test failures for nested aggregates.  I am going to debug that next.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/48


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5005 8f7bc0f5-282e-0410-95e3-8d296e9bb460