Workaround for [1].  There are three places where we use Logger.setLevel(). Of these, the most critical is in Banner where it is used to raise the log level for com.bigdata to WARN if it has not been explicitly set.  This is vital because logging at the default log level causes a tremendous performance hit due to conditional logging in the B+Tree and other critical packages.
[1] documents an issue where the SLF4J bridge does not implement Logger.setLevel().  As a workaround, Banner.banner() will the following message log @ ERROR if the com.bigdata log level has not been explicitly configured:

Unable to raise the default log level to WARN. Logging is NOT properly configured. Severe performance penalty will result.

The other two locations where Logger.setLevel() is used are not critical as they only provide for additional debugging information rather than providing a precondition for satisfactory performance.  They are DumpIndexSegment and CounterSetHTTPDServer.  In both cases, the (wrapped) NoSuchMethodException is trapped and an error is logged out.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/362 (SLF4J - LOG4J bridge does not implement setLevel).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5012 8f7bc0f5-282e-0410-95e3-8d296e9bb460