Per [1], the code path which replaces the graph variable with a constant corresponding to the sole known binding has been effectively disabled by comparing DataSetSummary.nknown with -1 rather than with 1.
We currently lack a means to convey a binding for the graph variable without rewriting the predicate such that the graph variable is effectively removed from the query. However, I think that we might have a solution for that when we get into the RTO integration. In the meanwhile, the comparison with -1 ensures that we continue to produce the correct solutions. Enabling this code path in the future might
provide a slight performance gain.

I have added a unit test to TestNamedGraphs to protect against regression for this ticket.  The test fails if the code path is enabled and passes when the code path is disabled.

The changes have been committed to both the 1.0.0 maintenance branch and to the development branch.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/359

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5013 8f7bc0f5-282e-0410-95e3-8d296e9bb460