Bug fix for https://sourceforge.net/apps/trac/bigdata/ticket/361 (IRunningQuery does not terminate promptly).
I have reduced the log level of the RejectedExecutionException from ERROR to WARN.  This exception is not really an error per say, especially for a Journal.  In a cluster this situation might cause problems with the termination of a query.  Basically, the situation is that we are unable to notify a QueryEngine on a node that some evaluation pass of an operator for a query is finished. The QueryEngine serving as the controller for that query needs to observe those events to decide when the query is "done".  So, if one node shuts down (rather than dies) while a part of the query is running on that node, then the controller might not receive the notice.  Hence the log @ ERROR as a clue for a query which is not terminating properly.  Such queries can be cancelled on the controller using their queryId and QueryEngine.getRunningQuery() to obtain the IRunningQuery and then invoking cancel() on that.  We plan to add some administrative features to the NanoSparqlServer for this sort of thing (detecting and canceling long running queries).  However, in all likelihood such queries will have RMI errors which lead to termination since the peer QueryEngine is dead.
{{{
                try {
                    t.context.getExecutorService().execute(
                            new SendHaltMessageTask(getQueryController(), msg,
                                    ChunkedRunningQuery.this));
                } catch (RejectedExecutionException ex) {
                    // e.g., service is shutting down.
                    log.error("Could not send message: " + msg, ex);
                }
}}}

Anyway, that is not really the source of this problem.

One problem was in AbstractChunkedResolverator.ChunkConsumerTask.call().  The loop there was not terminating as quickly as it should when the sink is closed.  Fixing that now results in the RejectedExecutionException error coming up more quickly.  The modified loop looks like this -- it explicitly checks buffer.isOpen().
{{{
                // while buffer (aka sink) is open and source has more data.
                while (buffer.isOpen() && src.hasNext()) {

                    // fetch the next chunk (already available).
                    final E[] chunk = src.nextChunk();

                    if (!buffer.isOpen()) {
                        /*
                         * Asynchronous close of the sink. By checking
                         * buffer.isOpen() here and in the while() clause, we
                         * will notice a closed sink more rapidly and close
                         * the source in a more timely manner.
                         * 
                         * @see https://sourceforge.net/apps/trac/bigdata/ticket/361
                         */
                        break;
                    }
                    
                    final F[] converted = resolveChunk(chunk);
                    
                    assert converted.length == chunk.length;
                    
                    // Note: Throws BufferClosedException if closed.
                    buffer.add(converted);

                    nchunks++;
                    nelements += chunk.length;
                    
                    if (log.isDebugEnabled())
                        log.debug("nchunks=" + nchunks + ", chunkSize="
                                + chunk.length);

                }

                final long elapsed = (System.currentTimeMillis() - begin);
                    
                if (log.isInfoEnabled())
                    log.info("Finished: nchunks=" + nchunks + ", nelements="
                            + nelements + ", elapsed=" + elapsed
                            + "ms, sink.open" + buffer.isOpen());

                return nelements;
}}}

However, the real problem was in AbstractChunkedResolverator#close().  The modified code is:
{{{
    public void close() {

        if (log.isInfoEnabled())
            log.info("lastIndex=" + lastIndex + ", chunkSize="
                    + (chunk != null ? "" + chunk.length : "N/A"));

        /*
         * Explicitly close the source since we will not be reading anything
         * more from it.
         *
         * @see https://sourceforge.net/apps/trac/bigdata/ticket/361
         */
        src.close();

        /*
         * Close the sink as well. The thread draining the sink's iterator will
         * notice that it has been closed. It will still drain anything already
         * buffered in the sink, but then the iterator() will report that no
         * more data is available rather than blocking.
         */
        buffer.close();

        chunk = null;
        
    }
}}}

With that change the test no longer logs the RejectedExecutionException because the IRunningQuery was correctly terminated.

I have also modified TestTicket361 to fail if the query is not terminated using an assert on the #of running queries on the QueryEngine.  If you comment out the src.close() in AbstractChunkedResolverator#close(), you can see that the test fails.  TestTicket361 is now run from the quads test suite.

I also made some minor modifications to Sesame2BigdataIterator to ensure that it complains if next() is called after close().  However, this does not appear to have been related to the problem.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5020 8f7bc0f5-282e-0410-95e3-8d296e9bb460