If you modify com.bigdata.relation.accesspath.WrappedAsynchronousIterator#hasNext() as follows then the source iteration appears to be closed in a timely manner. {{{     public boolean hasNext() {
        if(open && src.hasNext())
            return true;

        /*
         * Explicit close so we close the source as well when this is exhausted.
         * 
         * @see https://sourceforge.net/apps/trac/bigdata/ticket/361
         */
        close();
        
        return false;

    } 
}}}

This seems to me indicative of a pattern which we should be applying to both our ICloseableIterator and the openrdf ClosableIteration.  We probably need to audit all implementations of both interfaces for this pattern.

This commit contains the results of that code audit.

See https://sourceforge.net/apps/trac/bigdata/ticket/361

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5021 8f7bc0f5-282e-0410-95e3-8d296e9bb460