Checkpoint on integration of the query engine with the memory manager.  I have defined a new IQueryContext interface which exposes the IMemoryManager instance for the query.  The IMemoryManager instance is lazily allocated.  The IQueryContext is now passed into PipelineOp.newStats(), which provides a hook for sharing the state of objects on the native heap.  The DISTINCT operator based on the HTree is now integrated with the memory manager via the QueryEngine using this hook.
Allocations on the native heap are released back to the direct buffer pool in the tear down for the query.  However, we need to actually hook the Future of the IRunningQuery for this purpose (TODO) since we MUST NOT return direct ByteBuffers to the pool while a thread may still have access to the buffer.  Doing do is a recipe for data corruption since the buffer will be reassigned to another process in short order.

I have started work on a generalized aggregation operator for native memory.  This would only be used if (a) the DISTINCT is used in an aggregate or there are dependencies in the select value expressions; and (b) the estimated cardinality and data scale of the solution multiset that will be fed to the aggregation operator is very large (this can be estimated using the RTO, once that is integrated).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5024 8f7bc0f5-282e-0410-95e3-8d296e9bb460