I've written a test suite for the MemorySortOp and generalize the code to handle the incremental materialization of solutions and computed value expressions.  AST2BOpUtility has been updated to reflect the changed annotations for the MemorySortOp, which now defines both a SORT_ORDER and a VALUE_COMPARATOR annotation.
There are some known issues with comparisons of inline and non-inline IVs which MikeP will be addressing shortly.  Basically, we need an INeedsMaterialization option for operators which always require the materialization of non-inline IVs.  ORDER_BY and aggregation both have this characteristic since they must run over all solutions at once and can not re-try solutions which have failed with a NotMaterializedException.  As an initial step in that direction, I have defined an IVComparator and a test suite for that class and also added a compareLiteral(IV,Literal) to IVUtility.  While this covers some cases, it does not cover all as demonstrated by TestIVComparator.

There is also an apparent bug in the QueryEngine dealing with its ability to guarantee single threaded evaluation for operators such as sort, or perhaps it is just a bug in the test suite.  Regardless, I plan to look into this today and also look further into atomic decision points for the release of native memory back to the direct buffer pools in the QueryEngine.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5025 8f7bc0f5-282e-0410-95e3-8d296e9bb460