Relayered the AST nodes to introduce an abstraction for nodes which may appear within a group node.
Added a node for named subqueries and another node for an "include" of the named subquery within a join group.

Added a node for "external" operators which can run within the query plan.  There is no support for this yet.  This is basically the concept of a pipelined version of an "ExternalSet" for use by an IV aware application.  Ideally, we would write a driver similar to SubqueryOp for this and give the application a narrow interface along which it could accept and return solutions.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5047 8f7bc0f5-282e-0410-95e3-8d296e9bb460