update: slf4j-api-1.6.1.jar (was 1.4.3) update: slf4j-jdk14-1.6.1.jar (was 1.4.3)
Fixed compile errors with ant build caused by the presence of the old
dependencies.

Migrated SparqlQueryTest changes from 2.4.2.  Propagated the new
checkOrder flag to our derived test suites.

The dataset tests still fail (if enabled).

The following provide the SPARQL test suites:

 - ManifestTest - SPARQL 1.0 manifest based test suite.
 - SPARQL11ManifestTest - SPARQL 1.1 manifest based test suite.
 - SPARQLQueryTest - Driver for SPARQL query tests.

There is also:

 - SPARQLSyntaxTest - I assume that this is a test suite for SPARQL
   syntax (no query evaluation).

Note: By adding the SPARQL 1.0 and SPARQL 1.1 into a parent test suite
the nesting of the tests has been changed.  This broke the logic to
filter out the dataset tests (filterOutDataSetTests()) and the logic
to run a single test (getSingleTest()).  Both methods were rewritten
to be recursive so they do not make assumptions about the nesting of
the test suites.  Those methods are defined in BigdataQueryTest.

We are still passing all the SPARQL 1.0 tests.  We are failing 25
(errors) plus 5 (failures) of the SPARQL 1.1 tests.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5053 8f7bc0f5-282e-0410-95e3-8d296e9bb460