I have updated the sparql.jjt file and the associated AST files to 2.4-dev.  The relevant files are now in their original openrdf packages (org.openrdf.query.parser.sparql and org.openrdf.query.parser.sparql.ast).  I have also imported org.openrdf.model.vocabulary.FN which openrdf is using the define some of the XPath URIs.
The FN update dependency could be removed through changes to point at the FunctionRegistry URIs instead.

IN and NOT-IN are not working.  JeenB is looking at [1] for this.  Looks like it won't be fixed until 2.5.1!

The TCK behavior is unchanged by this import.  

The AST conversion has added coveraged but the behavior is unchanged as judged by the test suite.

[1] http://www.openrdf.org/issues/browse/SES-818

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5079 8f7bc0f5-282e-0410-95e3-8d296e9bb460