I have written and integrated a batch resolution step which annotates the openrdf AST parse tree with BigdataValue objects whose IVs have been efficiently resolved against the database.  I had to modify the ASTIRI, ASTBlankNode, ASTRDFLiteral, ASTNumericLiteral, ASTTrue, and ASTFalse classes to extend a common ASTRDFValue class with a getter and setter for an RDF Value object.  That property is set to a BigdataValue by the BatchRDFValueResolver.
I have requested support for this feature in openrdf.  A copy of the email follows below.

Jeen,

Would it be possible for you to introduce an ASTRDFValue object with an application settable property for the org.openrdf.model.Value?  ASTIRI, ASTBlankNode, ASTRDFLiteral, ASTNumericLiteral, ASTTrue, and ASTFalse would then extend ASTRDFValue rather than SimpleNode.  A concrete instance of this trival class is inline below my signature.

We would use this for efficient batch resolution of ASTRDFValue objects against the database, which is a step that we would like to take during the parser phase for reasons which I will not bore you with.  Batch resolution of RDF Values is important on a cluster where the code would otherwise do a series of point lookups.  I have implemented this on our side and it all works as desired.

Thanks,
Bryan

/**
 * An abstract base class for AST objects modeling RDF Values.
 *
 * Note that I have defined a "getRDFValue()" method since there is otherwise a name collision
 * with ASTIRI.getValue().
 */
abstract public class ASTRDFValue extends SimpleNode {

    private Value value;
    
    /**
     * @param id
     */
    public ASTRDFValue(int id) {
        super(id);
    }

    /**
     * @param parser
     * @param id
     */
    public ASTRDFValue(SyntaxTreeBuilder parser, int id) {
        super(parser, id);
    }

    public Value getRDFValue() {
        
        return value;
        
    }

    public void setRDFValue(final Value value) {
        
        this.value = value;
        
    }
    
}

See https://sourceforge.net/apps/trac/bigdata/ticket/370 (Direct AST => AST translation).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5083 8f7bc0f5-282e-0410-95e3-8d296e9bb460