I've written an equivalent for Sesame's BindingAssigner which takes a variable bound to a constant in a source binding set and replaces the IValueExpression associated with that variable with a Constant(var,val).
The felt like an ugly hack. 

1. I had to be aware when writing the unit test that the VarNode objects which were going to be modified by the AST optimizer needed to be distinct from those which were used in the expected AST against which the behavior of the optimizer was tested.  

Here, I think that we would be better served by a copy-on-write model just like the rest of the BOps.  I've been talking over with Martyn what we might have to do to put together a stateful visitor pattern for the pre-order iterator so you could inspect your position on the bop tree during a traversal.  It would be even nicer to be able to register what pattern you were looking for, but I am interested in something with less development time right now.  I want to bounce a few more thoughts around with Martyn and see if we can come up with something simple and workable.

2. I had to modify VarNode to (a) allow the caller to set the IValueExpression; and (b) to inspect the IValueExpression in getVar() so it could handle the case where the value expression had been rewritten as an IConstant rather than a IVariable, tunneling through to Constant.Annotations.VAR.

Here, I think that we are being tripped up by the eager conversion of the AST into IValueExpressions.  Further, we are actually dropping the original information in the AST.  For example, in this case the name of the variable was only retained in the IValueExpression (IV land) and not in the AST.  So, when I replaced the IValueExpression with a Constant, the variable was no longer accessible as part of the AST without tunneling through to the annotation on the Constant.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5101 8f7bc0f5-282e-0410-95e3-8d296e9bb460