See [1,2].
Removed code related to the stack idiom in BOpUtility.  This was causing an extreme performance problem in com.bigdata.bop.rdf.TestBOpUtility, where the unit test duration went from 9 seconds to 7 minutes.  I also remove the INodeOrAttribute stuff and all of the integration points in BOPUtility as we were facing not only a performance penalty but also a functionality problem (the stack was not being correctly populated with all nodes which were visited and was never seeing the annotation names when it traversed into an annotation).  I have replaced this functionality with a replaceWith() and replaceAllWith() on ASTBase and with a mutation() method on ModifiableBOpBase.  If a change is made to the AST, the nodes now notice this and can invalidate their cached value expression.  As part of this, I also modified the TermNode hierarchy to make the Constant or Var a child rather than an annotation. This makes it much easier to do rewrites and completely avoids the need for invalidation on structural change for TermNode.  I have not yet looked at how FunctionNode is being handled, but I plan to do so in conjunction with the annotation of the lexicon namespace on the query so we can remove that metadata from the external request for the value expression for a given function node.

I added an ASTSubqueryVariableScopeRewrite.  This should fix the problem with bottom up evaluation semantics for subqueries, which is part of [1].  However, it is possible that a subquery will have NO variables which MUST be bound in the scope of the parent.  Such subqueries should be rewritten to use the named/include pattern.  I have not yet done this, but the issue is documented on DefaultOptimizerList.

I have incorporated mroy's fix to the ASTNamedSubqueryOptimizer.  He modified it to recurse into subqueries looking for includes and not just the top level query.

I wrote a utility class diff(bop,bop) which reports the place where two BOp trees differ.  This made it possible to track down a problem where a Constant was given a Constant as an argument.  An exception is now thrown for this case.

More cleanup of the AST dump.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/232 (bottom-up evaluation).
[2] https://sourceforge.net/apps/trac/bigdata/ticket/370 (AST integration)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5121 8f7bc0f5-282e-0410-95e3-8d296e9bb460