We handle the subquery join in at three different ways.
1. SubqueryOp - Pipelined evaluation of a subquery in which the
   bindings flow through from the parent and back up into the parent.
   This is amenable to the push()/pop() pattern.

2. SubqueryHashJoinOp - The source solutions are fully materialized in
   a (JVM) hash table.  The subquery is run once.  Since the subquery
   is evaluated without any bindings from the parent, we do not need
   to use a push/pop pattern for this operator.

3. NamedSubqueryIncludeOp - The subquery is run once and its results
   are materialized in an HTree.  When the subquery is evaluated,
   there are no source bindings available.  The subquery solution set
   should contain bindings only for variables which were projected
   from the named subquery.  When we run the INCLUDE, we do not need
   to push/pop to change the lexical scope as the subquery solution
   set has already been computed.  We just do the hash join.

Thus, only pipelined subqueries need to use the push/pop pattern to
model the change in the visibility of the non-projected variables.

Deprecated and remove the use of the ASTSubqueryVariableScopeRewrite
class.  Wrote a SubqueryScopeOp which is used to apply the push() /
pop() pattern around a SPARQL 1.1 style subquery in order to hide
variables in the parent's scope which are not projected out of the
subquery.  Modified AST2BOpUtility to use this pattern for SPARQL 1.1
pipelined subquery evaluation.

Modified TestSliceOp to bring it into alignment with Matt's change to
permit 0L LIMIT.

Bug fixes to the test suite for the CONSTRUCT optimizer.  It was not
testing what it was supposed to be testing.

Added Comparator to AssignmentNode to support sorting of the
projection in order to make it easier to write some unit tests,
including the one for the CONSTRUCT optimizer.

Modified 3 of the aggregation unit tests to be consistent with Matt's
change to have COUNT() return an xsd:integer rather than an xsd:long.
Modified COUNT() to use the more efficient BigInteger.valueOf(long)
method.

Modified the HashJoinUtility to use the IBuffer abstraction for its
method arguments rather than a specific IBuffer implementation class.

Added a method to set all the children on the ModifiableBOpBase. This
was written to support the sorting of the children of the projection
node.

@see https://sourceforge.net/apps/trac/bigdata/ticket/232
@see https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5125 8f7bc0f5-282e-0410-95e3-8d296e9bb460