Modified xpath:substring function to confirm with the xpath specification.  However, the sesame 2.5 test suite has incorrect expected behavior for xpath:substring.  It appears that this is being fixed in 2.5-dev.  See [1]
Modified AbstractLiteralBOp to encapsulate the vf (BigdataValueFactory) field.  The vf field was not marked as [volatile], but it MUST be volatile to use the double checked locking pattern.  It is now private and volatile and access to the value factory is now always through getValueFactory() which uses the double checked locking pattern.  This makes the _get(IBindingSet) method a somewhat useless hangover.

Adjust some function URIs to correspond with those in use by Sesame's FN class.

[1] http://www.openrdf.org/issues/browse/SES-830

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5142 8f7bc0f5-282e-0410-95e3-8d296e9bb460