Bug fix to named graphs in the dataset declaration.  The default sesame dataset declaration processor was not using the versions of the URIs on which the IVs had already been resolved.
Added support for timeout (milliseconds) to QueryBase.

Added support for includeInferred to QueryBase and brought across support for this in AST2BOpUtility#toPredicate().  Also, some of the local access path logic had been commented out there and is now restored.  The method is now identical in semantics to the same named method in BigdataEvaluationStrategy3Impl.

https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5148 8f7bc0f5-282e-0410-95e3-8d296e9bb460