Fixed deadline support.  The deadline was being set on the AbstractRunningQuery rather than the QueryState.
Modified the NanoSparqlServer to use native evaluation.  This is a configuration option.  It defaults to true.  The test suites (single server) pass with that setting.  The default is in the ConfigParams class.

Modified the BigdataSailQuery impls to handle setDataset() using batch resolution of Values to IVs.

Modified the BigdataSailQuery impls to report the AST and optimized AST for EXPLAIN and STATUS.

https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5154 8f7bc0f5-282e-0410-95e3-8d296e9bb460