Rewrote the ASTEmptyGroupOptimizer and added some unit tests.  It was identifying the situations in which it should perform a rewrite but it was failing to effect the rewrite.
Modified the ASTSearchOptimizer to recognize magic search predicates within the main WHERE clause as well as within named subqueries.

Added an AST optimizer to lift out ServiceNodes which appear in contexts in which they would be evaluated more than once.  Such ServiceNodes are now lifted into a named subquery and replaced by the include of the named subquery solution set.

JoinGroupNode.toString(int) was not displaying the graph context. This was fixed by testing for that class in GroupNodeBase and explicitly incorporating that information into the rendered AST string.

Cleaned up toString(int) for NamedSubqueriesNode and NamedSubqueryRoot.

@see https://sourceforge.net/apps/trac/bigdata/ticket/267

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5175 8f7bc0f5-282e-0410-95e3-8d296e9bb460