Modified JoinGroupNode to report out maybe bound + bound variables.
Modified TermId to show the cached value and fixed fromString() to
still decode correctly.

Modified BOpUtility#toString() to show recursive structure for
annotations which are PipelineBOps.  This makes the subqueries visible
in a query plan.

Bug fix for running multiple named subqueries in parallel. The
NamedSubqueryOp now correctly uses the IMemoryManager associated with
the root IRunningQuery.

Added static analysis methods to GraphPatternGroup based on those
already defined in JoinGroupNode.  This set of methods supports the
"must" bound incoming, the "must" bound in the group (and optionally
recursively for the child groups), the "maybe" bound by the group (and
optionally recursively for the child groups).  Unions are handled for
both cases which look down the tree recursively.  I also added a
method to report the "must" bound variables on QueryBase.  There is
now a test suite for all this stuff at the AST level.

The TestStaticAnalysisMethods suite is now developed enough to clearly
indicate why certain "bottom-up" evaluation queries in the TCK are
failing.

Modified the type for IGroupMemberNode#getParent() by filling in the
generic type parameter.

Added IGroupMemberNode#getParentJoinGroup() and
IGroupMemberNode#getParentGraphPatternGroup().

Static analysis of join variables for hash tables is not as easy as I
had hoped.  It depends on the expected cardinality of the named
solution set and the expected cost of evaluating some of the joins in
group in which that solution is included before / after the join
against the named solution set.

In the case where the INCLUDE is embedded into a child group, we
currently *assume* that some things are already bound by the parent
group.  However, this really depends on running the INCLUDE after the
non-optional joins in the parent, which is itself just a heuristic. In
fact, the entire notion of GraphPatternGroup#getIncomingBindings()
depends on this heuristic!  Truely dynamic analysis is much more
complicated.

@see https://sourceforge.net/apps/trac/bigdata/ticket/370
@see https://sourceforge.net/apps/trac/bigdata/ticket/368

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5180 8f7bc0f5-282e-0410-95e3-8d296e9bb460