Bug fix to TermId#equals().  It was comparing the cached value with hasValue() (a boolean) for the other TermId.
Test suite change for TermId now that two TermIds without cached values which are NULL IVs are not equals.  The test suite now verifies that they have the same data and that equals() reports false.

Bug fix to test suite for parse tree to AST translation handles case where the parse tree is not available gracefully.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5206 8f7bc0f5-282e-0410-95e3-8d296e9bb460