A little more work on the simple optional optimization.  I have not been able to make the materialization pipeline work correctly when lifting a filter which had materialization requirements which could be satisfied against the required joins in the parent group.  I am also unsure whether this is worth the effort, as the variables to be materialized for the optional join probably have to be "ALWAYS" since we can not run the optional join multiple times in the parent.
I'd like to talk this over with Mike before proceeding on it. The logic to handle the lift of filters which do not require materialization is still in place but the rest of the logic is conditionally disabled in ASTSimpleOptionalOptimizer.

Some changes to StaticAnalysis in support of optional statement pattern nodes in join groups.

@see https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5208 8f7bc0f5-282e-0410-95e3-8d296e9bb460