Working on the ASTGraphGroupOptimizer.  This applies inherited constraints during top-down traversal.  If a constraint is clearly violated, e.g., GRAPH uri1 {... GRAPH uri2 { ... } ... } then that is reported as a translation time error.  I have not written any new tests for this yet, but we have some coverage via the existing TestNamedGraphs in the bigdata-sails module.
The following two patterns are more complicated as I need to generate
statement patterns which get recognized and translated efficiently.
Sometimes into a distinct term advancer.  Sometimes into an inline
access path.  I can't just add an IElementFilter to the statement
pattern since it would be too difficult to recognize.  To support
this, I am thinking of adding explicit annotations for DISTINCT and
EXISTS onto StatementPatternNode for this stuff. That way we can do
the right thing in toPredicate().

Semantics for GRAPH ?g {} (and unit test).

Semantics for GRAPH <uri> {} (and unit test).
	  
We may require an inline access path attached to a predicate in order
to support GRAPH ?g {}, which is basically an inline AP for the
dataset.  That should probably be another annotation on the
StatementPatternNode, perhaps COLUMN_PROJECTION (var,vals) or
IN(var,vals)?

@see https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5212 8f7bc0f5-282e-0410-95e3-8d296e9bb460