Bug fix and optimization for ASTGraphGroupOptimizer.  If the same graph variable is used in the outer and inner groups then we do not add the SameTerm constraint.  If we do add the SameTerm constraint, then toVE() is now invoked to set the value expression.
Modified the SPARQL syntax test suite to now run against our version of the SPARQL parser.

Javadoc and cleanup on the materialization pipeline code.  It is being centralized in AST2BOpUtility for the moment. There was duplication in Rule2BOpUtility.

Moved the logic for computing materialization requirements into StaticAnalysis. It was used by both ComputedMaterializationRequirement and Rule2BOpUtility and maybe AST2BOpUtility.  It is now centralized.

Added TCK test for construct with reification.  This test is failing.  The root cause (or at least the initial root cause) is RDF Collection handling in the parse tree => AST translation.

Added test suite for [] blank node syntax for the parse tree => AST translation.

Identified and reported a bug in base URI resolution for Sesame 2.5.0 (and probably earlier versions).  The trailing path component is being stripped, not just the URI fragment.  We now have tests for this, but they are conditionally disabled until Sesame provides a fix.

Added/updated test suite for RDF Collections handling in the parse tree => AST translation.  These tests are still failing, but they at least model the expected behavior.  I will need to dig into the sparql.jjt grammar file and visitors to figure out what's going on here.  The problem is doubtless that there are interactions with the SPARQL 1.1 property path grammar productions, which are currently commented out.

@see https://sourceforge.net/apps/trac/bigdata/ticket/370

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5214 8f7bc0f5-282e-0410-95e3-8d296e9bb460