A variety of license / copyright notice cleanup related changes.
done. Added com.bigdata.Depends.  This file has the name of each dependency component, the project URL, and the license URL.  This can be used to self-report on the dependencies.  Not all dependencies are required for all deployments, but this class does not presently allow selective reporting based on the kind of deployment (standalone versus cluster).

done. Modified the Banner class to include a notice for each dependency, including the name of the component, the project URL, and a link to the license file.

done. Our own license file is not making it into the top-level of bigdata.jar or bigdata.war. The JAR now includes the NOTICE files that we have in our source tree for the code that is bundled into the JAR. The JAR does not bundle those dependencies so it does not need to include the LEGAL/* files for them dependencies. However, we DO need the LICENSE files for Apache and Sesame since we have imported some code from those projects. The WAR actually redistributes dependencies. It needs to include all the files in the LEGAL directories in our source tree (these are the dependency license files) plus all the NOTICE files in those dependencies.  Wrote an ant task to concatenate all NOTICE(.txt) files into a single combined NOTICE file. The JAR now includes the combined NOTICE for the bigdata source tree as bundled into the JAR.  The WAR now includes the combined NOTICE for all dependencies (even those which are not actually bundled into the
WAR).

done. The overview.html file is missing for the javadoc.

done. Verify WAR deployment.

done. Bump up the branching factor for the SPO relation for the default kb when deploying under the webapp.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5230 8f7bc0f5-282e-0410-95e3-8d296e9bb460