Working on support for inline access paths.  As of this commit, joins against the lexicon or statement indices operate on IBindingSets read from the index rather than IElements.
Tracked down the problem mentioned above.  It turns out the Predicate#asBound() was not associating the IVariable with the Constant on the new predicate.  The new path for bind() was therefore failing to propagate bindings for the variable for the asBound predicate.  This resulted in NotMaterializedExceptions because the lexicon join was failing to associate the resolved Value with the variable because its "variable" quality was lost by asBound().

The change was to Predicate#asBound() on line 360.

Old:
{{{
            tmp._set(i, val.clone());
}}}
New:
{{{
            tmp._set(i, new Constant(var, val.get()));
}}}

Also fixed a bug in the BOpContext#bind() variant for access paths visiting solutions where it was failing to propagate the IV's cached value.  Added a test suite for this bind() method.

The new code path in PipelineJoin to use AccessPath#solutions() rather than AccessPath#iterator() is enabled in this commit, but can be disabled by changing the following code at line 1640.
{{{
				    if(true && accessPath instanceof IBindingSetAccessPath) {
				        // Handle join in terms of an IBindingSet iterator.
                        handleJoin2();
				    } else {
				        // Handle join in terms of an IElement[] iterator.
				        handleJoin();
				    }
}}}

@see https://sourceforge.net/apps/trac/bigdata/ticket/209

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5269 8f7bc0f5-282e-0410-95e3-8d296e9bb460