The scale-out data loader should now support both TermIds and BlobIVs.  I've added unit tests for quads and with blobs.  A split handler is now registered in scale-out for the blobs index.  One of the quads mode tests is failing, but it is failing because the [c] position is not always bound on a statement in quads mode.  I am still looking at this issue.  I think that we will probably wind up adding asserts for the key arity into SPOIndexWriter and then fixing the DataLoader and AsynchronousStatementBufferFactory to set the [c] position to the resource from which the data are being loaded (default context?).
I believe that I have found the cause of the errors in TestRollbacks during CI.  I think that there has been a side-effect through the private fields on the BigdataValueFactoryImpl.  If the same namespace is used in CI for different triple store instances, then we can have old data in those fields.  Perhaps we should explicitly (and atomically) clear the namespace in CI before a test is run?  (I ran into a similar problem in the IV unit tests which is how I finally figured out where the side-effect was coming from.)

Javadoc on TestSubQuery which captures the dialog with Matt.  I have not yet fixed the bug.

I still need to update the scale-out data loader to write on the full text index.

I still need to provision the BLOBS index defaults for high efficiency in scale-out.

I am committing now to get feedback on these changes from CI.  I hope to load BSBM in scale-out tomorrow to verify the blobs handling on a cluster, at which point I should be able to close off this set of issues and begin to focus on query optimizations, inline APs, and the RTO integration.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5287 8f7bc0f5-282e-0410-95e3-8d296e9bb460