Added an AST optimizer to flatten nested unions.  There is one unit test based on BSBM Q6.  There is another unit test for the correct rejection of ASTs where we can not flatten the union.
Added concrete instances of each of the BSBM BI queries along with the PC 10 data set and a harness which may be used to examine the behavior of the queries.  The harness currently loads the data set for each query presented, which makes it a bit slow.  There is an emptyDataset file if you just want to see how the queries are being translated and a PC 10 dataset if you want to get a very rough sense of the running time for the queries.

Query6 had nested unions which were causing a problem.  I've added the ASTFlattenUnionsOptimizer described above to address this.  However, Query6 is still very slow because we are running the subquery "as-bound" rather than first.  This can be verified by comparing the performance with query6b, which is just a rewrite of query6 using a named subquery for the subselect.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5302 8f7bc0f5-282e-0410-95e3-8d296e9bb460