Clarified the javadoc in Steps that this is for sequential execution of subqueries which have side-effects (e.g., the fast closure rules). The Union operator does the same thing but allows the subqueries to be executed in parallel.  Modified AST2BOpUtility to use Union rather than Steps when executing named subqueries in parallel.
Javadoc in ASTSparql11SubqueryOptimizer. This is capable of lifting all Sub-Selects into named subqueries (conditional compilation - change a constant).  However, I think that what we really want to be doing here is either chunking the subquery (a set of source solutions) or feeding in ALL source solutions from the parent group.

@see https://sourceforge.net/apps/trac/bigdata/ticket/377

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5303 8f7bc0f5-282e-0410-95e3-8d296e9bb460