Updated the wiki page for the NanoSparqlServer REST API [1] to reflect the new DELETE by access path method.
I've added a delete by access path method into the DeleteServlet, which was easy enough. 

The hard part is the correct encode/decode of the URIs and Literals so they can be sent through.  For example, a Literal with an embedded quotation mark or a URI with an embedded GT or LT symbol.  The most general solution for this is to handle the SPARQL code point sequences [2], which is a bit of work (in fact, Sesame only handles one of the two code point sequence forms - it seems that JavaCC does not recognize the 8 character code point sequences.  For the moment, the implementation ignores this issue.  The relevant code is in EncodeDecodeValue and can be enhanced once the API change is proven useful.  The most critical problem that someone is likely to encounter is a quotation mark embedded within a Literal.  At present, that will not come across the API correctly.  (It will be reported as a bad request exception at the HTTP layer.)  There are two unit tests which demonstrate the encode/decode problem for URIs and Literals.

Committed revision rXXXX.  (Committed against both branches/BIGDATA_RELEASE_1_0_0, which is the 1.0.x maintenance branch, and TERMS_REFACTOR_BRANCH, which is the basis for our 1.1.x release.)

[1] https://sourceforge.net/apps/mediawiki/bigdata/index.php?title=NanoSparqlServer#DELETE
[2] http://www.w3.org/TR/sparql11-query/#codepointEscape

@see https://sourceforge.net/apps/trac/bigdata/ticket/392

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5332 8f7bc0f5-282e-0410-95e3-8d296e9bb460