I have added a "context-uri" request parameter which may be used with the INSERT and UPDATE operations in the REST API.  This request parameter is documented at [1] and only effects operations against a quads mode database.
Note: When present, the "context-uri" parameter overrides the default context for all resources which are being inserted.

Note: There is also a bug fix.  The INSERT with URI(s) method was failing to impose the URI of the resource being loaded as the default context.  It now uses the URI of the resource that was loaded as the default context unless the "context-uri" request parameter is also specified.

[1] https://sourceforge.net/apps/mediawiki/bigdata/index.php?title=NanoSparqlServer

@see https://sourceforge.net/apps/trac/bigdata/ticket/393

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5333 8f7bc0f5-282e-0410-95e3-8d296e9bb460