I have stubbed out the code to run a Sub-Select using the hash index + hash join pattern.  The new code path does not quite work yet, but it is close.
It appears that we are failing to lift a SubqueryRoot into the parent join group. It winds up in a join group by itself, which means that we have two levels of indirection when running a subquery, which is inefficient.

I have refactored a boolean for enabling or disabling the Htree based hash joins into AST2BOpContext.  The HTree based hash joins still have the de-serialization bottleneck so they are disabled by default.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5346 8f7bc0f5-282e-0410-95e3-8d296e9bb460