Added a JVM version of the NamedSubqueryOp and refactored AST2BOpUtility and AST2BOpContext to isolate the boolean switches for generating the hash-join pattern for sub-group evaluation and the hash-join pattern for sub-query evaluation.  The JVM version of the NamedSubqueryOp will be used unless native hash joins are specified in AST2BOpContext.
NamedSubqueryIncludeOp is no longer used.  It has been replaced with the more general purpose SolutionSetHashJoinOp (HTree) and JVMSolutionSetHashJoinOp.  The class is still hanging around as it declares some annotations used by the other hash join operator classes but it will be eventually removed entirely.

Modified TestASTLiftPreFiltersOptimizer to succeed and log an error indicating that no tests have been written.  We still need to write this ast optimizer.  It is responsible for lifting pre-filters into the parent group.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5350 8f7bc0f5-282e-0410-95e3-8d296e9bb460