Added a Distinct IV filter based on the HTree class and a simple unit test.
The current filter pattern lacks a means to communicate the IRunningQuery to the filter (other than perhaps via the context reference, but no such pattern has been standardized).  This means that the HTreeDistinctFilter manages its own MemoryManager instance rather than piggybacking on the one associated with the running query.  That means that the HTreeDistinctFilter will eat a minimum of 1M of native RAM when it runs.

The filter is hooked into AST2BOpBase, but it is disabled in the code using an if-then-else block with a constant.  There are inline comments about the minimum memory overhead of using this operator.

The HTreeDistinctFilter uses encoded IVs and thus serves as a handy proof of concept for the use of encoded IVs in the HTree based hash joins.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5372 8f7bc0f5-282e-0410-95e3-8d296e9bb460