Rolled back StaticAnalysis to r5421; Removed the delta from r5409 and instead modified getDefinitelyIncomingBindings() to NOT recurse into an optional group.  This fixes the bottom up query evaluation problems.
Removed the unit test for ASTSimpleOptionalOptimizer for lifting a filter for the case which I was not able to make work correctly.  Complex optional groups are now evaluated efficiently so we no longer have any reason to make the logic for simple optionals more complicated.

Bug fix to TestASTNamedSubqueryOptimizer.  It was trying to compare an IVariable[] to a VarNode[].  This test still fails due to a fail() call.  I have left that in to fix the test.  It needs to be rewritten to build up the AST directly rather than parsing  SPARQL query and then running all the optimizers.  The way it is written right now we do not have sufficient control to guarantee that the test tests what it should test.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5424 8f7bc0f5-282e-0410-95e3-8d296e9bb460