Bug fix to the query plan generator for LIMIT.
SliceOp was being added to the query plan as a non-pipelined (at-once) operator.  The query engine was waiting for all source solutions to be materialized before it ran the slice.  Hence the entire query was running to completion before the SLICE was imposed.

I've modified AST2BOpUtility to specify pipelined semantics for the SliceOp.  I modified SliceOp to refuse construction as a non-pipelined operator and verified that this does not break anything in the test suites.

A "SELECT * {?s ?p ?o} LIMIT 1" query now finishes immediately.  I have added this as a query to the CI/govtrack query set so we can detect if SLICE breaks.

Organized imports in IVariableOrConstant.

BlockingBufferWithStats could potentially increment the #of units out when the output buffer was full.  This has been fixed.

@see https://sourceforge.net/apps/trac/bigdata/ticket/400

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5448 8f7bc0f5-282e-0410-95e3-8d296e9bb460