hashjoin02 was moved into a different directory.
QueryHints#HASH_JOIN is now on AST2BOpBase.

TestBOps was not verifying that the hash join was performed nor that the query hint was interpreted correctly.  Both are done by TestHashJoins so I removed the corresponding test from TestBOps.

Finished a unit test of the query hint optimizer.

Bug fix to the query hint optimizer when context.queryHints is null (it should never be null when we are running queries through the system, but it was null in the unit test which was hooked in at a different level).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5542 8f7bc0f5-282e-0410-95e3-8d296e9bb460