Optimized version of native DISTINCT filter for solution sets.  This is based on the HTreeHashJoinUtility.  This can be exercised with CI/govtrack/distinct01.rq.  [There is still more work to be done to vector acceptSolutions() and filterSolutions().  Also, the JVM hash utility class needs to be updated and the test suite for htree based DISTINCT solution sets is broken and needs to be migrated to an RDF specific data set.]
Added logic to handle materialization of projections when a query does
not use a SLICE.  Materialization done from within the query plan
needs to occur before the SLICE operator since the SLICE will
interrupt the query evaluation when it is satisfied, which means that
downstream operators will be canceled. Therefore a materialization
operator can not appear after a SLICE.  However, doing materialization
within the query plan is not efficient when the query uses a SLICE
since we will materialize too many solutions.  This is true whether
the OFFSET and/or the LIMIT was specified.

Moving the materialization for the projection inside of the query plan
uncovered a bug where materialization within an OPTIONAL join group
was treated as proof that the variable was materialized.  However,
this is not always true.  For example, in a complex optional group a
variable which is known bound in the parent may be filtered by a JOIN
before a FILTER is applied which requires us to materialize that
variable.  If the materialization step is added after the JOIN within
the complex optional group, then some values of the known bound
variable will not be bound.  Either we need to move the
materialization step outside of the optional group such that all
values are known bound or we must not add variables which are
materialized within an optional group to the set of known materialized
variables. [This bug has not been fixed yet.]


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5562 8f7bc0f5-282e-0410-95e3-8d296e9bb460