Bug fix to LexPredicate.  Modifications to clone() and asBound() parallel to those made to Predicate and SPOPredicate.
Modified UniquenessFilter to use a HashSet.  This is faster than LinkedHashSet when order does not matter and you are not using the set iterator.

Change to BOpUtility#getDistinctArgumentVariables() return value and TODO for optimization.  Change to AccessPath#solutions(), which is the only place this method is used.  It is a bit of hot spot, especially when there are fully bound AP tests, and there is a HashSet (UniquenessFilter) allocation for each invocation.

Optimizations to comparison of annotation maps.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5680 8f7bc0f5-282e-0410-95e3-8d296e9bb460