I have added query hints to control the sampling limit, the default behavior when sampling is disabled, for the SCAN + FILTER versus PARALLEL SUBQUERY decision for named and default graph queries.  See QueryHints#ACCESS_PATH_SAMPLE_LIMIT and QueryHints#ACCESS_PATH_SCAN_AND_FILTER.  These things may also be overridden via query hints either for the entire query (scope=Query) or for specific statement patterns (when the scope is more selective).
I have not changed the default behavior.  The sample limit is still 100 and the behavior if you disable sampling will default to SCAN+FILTER.  I want to look at integrating the DataSetJoin into the default graph code path next (it is already used in the named graph code path to handle the PARALLEL SUBQUERY case).

@see https://sourceforge.net/apps/trac/bigdata/ticket/407

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5697 8f7bc0f5-282e-0410-95e3-8d296e9bb460