Checkpoint.  I have removed several classes which were holdovers from the old SPARQL parser integration.  I have added support in the AST for MINUS, which is now modeled as a JoinGroupNode with the MINUS annotation.  I have not yet revisited the AST model for EXISTS or NOT EXISTS, but I did create a variant on the SPARQL => AST unit tests for EXISTS parallel to the existing test for NOT EXISTS.  I also added a unit test for MINUS in the SPARQL => AST test suite.
The changes to the query plan generator MINUS and (NOT) EXISTS will interact with the decisions about what variables have been materialized.  Materialization within the scope of the child join group (for MINUS) or the filter (EXISTS or NOT EXISTS) must not be visible outside of that scope.  I have not yet captured this behavior in the code.

My next step will be to replace JoinAnnotations#OPTIONAL with a JOIN_TYPE annotation. And then verify that the existing hash join operators can be used with EXISTS and NOT EXISTS style joins per HashJoinEnum (which will be renamed to JoinTypeEnum).

@see https://sourceforge.net/apps/trac/bigdata/ticket/414

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5764 8f7bc0f5-282e-0410-95e3-8d296e9bb460