Various changes in support of [1,2,3].
All EXISTS, NOT EXISTS, and MINUS tests now pass.  The last knarly bit was getting the FILTER variable scope right for an inner FILTER inside of an EXISTS.  I did this by writing StaticAnalyis#findParent() and then using that to figure out when the first GraphPatternGroup parent of a JoinGroupNode was a FilterNode, in which case the join group node is clearly appearing inside of an (NOT) EXISTS FILTER.  There is still an open issue with the dynamic ordering of MINUS, which needs to be based on the same principle by which filter variables govern when a filter can run.  The EXISTS and NOT EXISTS filters now run using the (much) faster hash join pattern rather than the SubqueryOp.

The issue [2] with following renames of projected variables has been fully resolved.

Only minor steps have been taken towards [3] (paying attention to exogenous bindings).

[1] https://sourceforge.net/apps/trac/bigdata/ticket/414 (SPARQL 1.1 EXISTS, NOT EXISTS, and MINUS)
[2] https://sourceforge.net/apps/trac/bigdata/ticket/430 (StaticAnalysis does not follow renames of projected variables)
[3] https://sourceforge.net/apps/trac/bigdata/ticket/412 (StaticAnalysis#getDefinitelyBound() ignores exogenous variables.)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5773 8f7bc0f5-282e-0410-95e3-8d296e9bb460