Updated release notes.
The 1.0.x branch actually has two versions of RWStore.properties in the bigdata-war module.  The one in bigdata-war/src/resources is the one that winds up being installed in the WAR.  I have therefore made it a copy of the one in bigdata-war/RWStore.properties.  This issue was already resolved in the 1.1.x branch (TERMS_REFACTOR_BRANCH).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5816 8f7bc0f5-282e-0410-95e3-8d296e9bb460