Removed IVCache#dropValue() from the code base. It was unused and implies a pattern which can not be made to work.
Added IVCache#clone(boolean).  This is in support of [1].  The AbstractIV#cache field was providing a hard reference from the key of the termCache weak value map to the value of that map. That caused the values of the map to become pinned and is a root cause memory leak.  This was identified based on a heap dump after running the BSBM explore use case ~ 80 times (until the memory leak began to degrade the ability of the OS to buffer the disk as evidenced by increasing IO Wait).

Modified the IV test suites per above.

Abstracted the termsCache using an ITermCache interface and flyweight TermCache wrapper class which delegates the necessary methods through to the backing concurrent weak value cache impl.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/437 (Thread-local cache combined with unbounded thread pools causes effective memory leak)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5829 8f7bc0f5-282e-0410-95e3-8d296e9bb460